# docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateCommentInput:
    properties:
      content:
        type: string
    type: object
  handlers.CreatePostInput:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  handlers.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  handlers.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  handlers.UpdatePostInput:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
  handlers.UpdateUserInput:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
  models.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: dsaputra5403@gmail.com
    name: Dimas Saputra
  description: Ini adalah server API untuk GopherPost
  termsOfService: http://swagger.io/terms/
  title: GopherPost API
  version: 1.0.0
paths:
  /api/comments/{id}:
    delete:
      description: Menghapus comment permanen berdasarkan ID
      parameters:
      - description: ID Komentar (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus komentar
      tags:
      - comments
  /api/posts:
    post:
      consumes:
      - application/json
      description: Membuat post dengan judul dan konten. Butuh token JWT.
      parameters:
      - description: Data Post
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePostInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Membuat postingan baru
      tags:
      - posts
  /api/posts/{id}:
    delete:
      description: Menghapus post permanen berdasarkan ID
      parameters:
      - description: ID Postingan (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus postingan
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Mengubah judul atau konten post berdasarkan ID
      parameters:
      - description: ID Postingan (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Data Update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePostInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Edit postingan
      tags:
      - posts
  /api/posts/{id}/comments:
    post:
      consumes:
      - application/json
      description: Memberikan komentar pada postingan tertentu
      parameters:
      - description: ID Postingan (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Isi Komentar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCommentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Kirim komentar
      tags:
      - comments
  /login:
    post:
      consumes:
      - application/json
      description: Tukar email dan password dengan Token JWT
      parameters:
      - description: Kredensial Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Masuk ke aplikasi
      tags:
      - auth
  /posts:
    get:
      description: Mengambil daftar postingan dengan pagination
      parameters:
      - description: Nomer Halaman
        in: query
        name: page
        type: integer
      - description: Jumlah Data per Halaman
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Melihat semua postingan
      tags:
      - posts
  /posts/{id}:
    get:
      description: Mengambil detail post berdasarkan ID
      parameters:
      - description: ID Postingan (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Melihat satu postingan
      tags:
      - posts
  /posts/{id}/comments:
    get:
      description: Mengambil semua komentar berdasarkan ID Postingan
      parameters:
      - description: ID Postingan (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Dapatkan komentar
      tags:
      - comments
  /register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan akun baru ke sistem
      parameters:
      - description: Data User
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Daftar user baru
      tags:
      - users
  /users:
    get:
      description: Retrieves a list of all registered users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all users
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletes a user identified by ID.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      description: Mengambil data detail user berdasarkan ID (UUID)
      parameters:
      - description: ID User (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Lihat profil user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates the name and email of a user identified by ID.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update user profile
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"